refactor: extract components from useAiChat and RPG types
- Extract modelStorage.ts for AI provider/model storage utilities
- Extract ttsHandlers.ts for TTS functionality
- Extract useMediaGenerators.ts for image/video/music generation
- Update useAiChat.ts to use extracted utilities
- Split RPG types/index.ts into modular files:
  - character.ts, combat.ts, item.ts, skill.ts
  - map.ts, shop.ts, compendium.ts, events.ts, constants.ts
- Update types/index.ts to re-export from modular files
